perf: optimize memory usage with lazy DataSource init and per-file DF-to-numpy conversion
Phase 1: Stage間データ解放
- Add StageDataConfig dataclass with factory callable for lazy DataSource initialization
- Refactor learn_multi_stage() to run each stage in isolated helper functions (_run_stage1, _run_stage2)
- Call gc.collect() after each stage to ensure prompt memory release
- Fix .datasource attribute bug (FileDataSourceSpliter has no .datasource property)

Phase 2: ページ単位の逐次DF→numpy変換
- Consolidate FileManager.__init__ read and convert loops into single per-file loop
- Extract _load_feather() helper method for cleaner file loading dispatch
- Change moveLabel intermediate array dtype from float32 to float16 in convert_preprocessing_df_to_numpy

+++ src/maou/domain/data/schema.py
@@ -1029,11 +1029,14 @@ def convert_preprocessing_df_to_numpy(
     )
 
     # 1D List column: moveLabel List[Float32] -> (N, MOVE_LABELS_NUM)
+    # Polarsの内部データはFloat32だが，structured arrayのtarget dtypeはfloat16．
+    # float16を指定することで_explode_list_column内のastype変換後の中間配列が
+    # float32の半分のサイズになり，定常メモリを削減する．
     array["moveLabel"] = _explode_list_column(
         df["moveLabel"],
         n,
         (MOVE_LABELS_NUM,),
-        np.dtype(np.float32),
+        np.dtype(np.float16),
         nest_depth=1,
     )
